refactor pipeline cache
